Backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
python-dotenv>=1.0
pydantic>=2.12
langgraph>=0.3
langchain>=1.0
langchain-core>=1.0
langchain-openai>=1.0
langchain-community>=0.4
langsmith>=0.7
pandas>=2.3
numpy>=2.0
yfinance>=0.2
ta>=0.11
scikit-learn>=1.7
statsmodels>=0.14
xgboost>=3.0
tensorflow>=2.21
prophet>=1.1
torch>=2.10
transformers>=4.40
safetensors>=0.4
